refactor: コンポーネントに切り出しました(後は任せました)
